Fix port retrieval from communication containers
